fixed and tested cleaning and updating code

+++ src/clean-airport-data.R
@@ -34,26 +34,35 @@ airports <- data.frame(airport_codes, airport_states)
 # Run
 # put the resulting xls file into ext-data
 
-airport_data <- read.xls("ext-data/APM-Report.xls")
-
-# remove unnecessary rows, and rename columns
-airport_data <- airport_data[-(1:6),1:4]
-
-names(airport_data) <- c("Facility", "Date", "Scheduled_Departures", "Scheduled_Arrivals")
-
-# get rid of silly lines at the end
-airport_data <- airport_data[airport_data$Facility %in% airport_codes,]
-
-airport_data$Date <- as.Date(as.character(airport_data$Date), format = "%m/%d/%y")
-airport_data$Scheduled_Arrivals <- as.numeric(as.character(airport_data$Scheduled_Arrivals))
-airport_data$Facility <- as.character(airport_data$Facility)
-
-# png("figures/airport-arrivals.png", width = 1000, height = 1000)
-# ggplot(airport_data, aes(x = as.Date(Date), y = Scheduled_Arrivals)) + geom_line() + facet_wrap(~Facility) + theme_bw()
-# dev.off()
-
-# ok, what we need to do is match these up to states
-airport_data_states <- merge(x = airport_data, y = airports, by.x = "Facility", by.y = "airport_codes")
-names(airport_data_states)[5] <- "State"
-
-write.csv(airport_data_states, "clean-data/airport_data.csv", row.names = FALSE)
+# check if the file exists
+f <- "ext-data/APM-Report.xls"
+if (file.exists(f)){
+  
+  airport_data <- read.xls(f)
+  
+  # remove unnecessary rows, and rename columns
+  airport_data <- airport_data[-(1:6),1:4]
+  
+  names(airport_data) <- c("Facility", "Date", "Scheduled_Departures", "Scheduled_Arrivals")
+  
+  # get rid of silly lines at the end
+  airport_data <- airport_data[airport_data$Facility %in% airport_codes,]
+  
+  airport_data$Date <- as.Date(as.character(airport_data$Date), format = "%m/%d/%y")
+  airport_data$Scheduled_Arrivals <- as.numeric(as.character(airport_data$Scheduled_Arrivals))
+  airport_data$Facility <- as.character(airport_data$Facility)
+  
+  # png("figures/airport-arrivals.png", width = 1000, height = 1000)
+  # ggplot(airport_data, aes(x = as.Date(Date), y = Scheduled_Arrivals)) + geom_line() + facet_wrap(~Facility) + theme_bw()
+  # dev.off()
+  
+  # ok, what we need to do is match these up to states
+  airport_data_states <- merge(x = airport_data, y = airports, by.x = "Facility", by.y = "airport_codes")
+  names(airport_data_states)[5] <- "State"
+  
+  write.csv(airport_data_states, "clean-data/airport_data.csv", row.names = FALSE)
+} else {
+  out_string <- "Airport data file not found, follow rake instructions to download it"
+  print(out_string)
+  write.csv(out_string, "clean-data/airport_data.csv", row.names = FALSE)
+}
